fix: Correct how errors are displayed in new.html
This fix makes sure the correct error message is displayed at the
top of the page and displays the error for the time_spent field
underneath the time_spent field in the form if the validation for that
field fails. "An integer greater than or equal to 1 is required" was
being displayed for any error at the top of the page instead of showing
the correct error.