feat: Allow users to edit journal entries
Renamed AddEntryForm to EntryForm since the form is used both to
create and edit journal entries.

Connected the edit entry link in the detail page.

Each field of the edit form is populated from the database.

Altered the detail view and the detail page so fields in the detail
page will now be populated from journal_post.attribute rather than
having to use a for loop.

Enabled CSRF protection globally per the documentation
https://flask-wtf.readthedocs.io/en/latest/csrf.html#setup which
is preventing a CSRF error from occurring when the fields in the
edit form are pre-populated.